fix: move yauzl.fromBuffer extraction into utils-sh, remove extract-zip entirely
Replace extract-zip with yauzl.fromBuffer() in @goldstack/utils-sh's unzip()
function, then have @goldstack/project-build delegate to it instead of
duplicating the extraction logic. This fixes the hanging extraction issue
in both project-build (template extraction) and project-repository
(project download/extraction) by avoiding yauzl's fd-slicer ref-count
leak and using buffer-based extraction (BufferSlicer) throughout.
