+++ workspaces/utils/packages/utils-sh/package.json
@@ -37,11 +37,11 @@
   "dependencies": {
     "@goldstack/utils-log": "0.3.34",
     "archiver": "^5.3.1",
-    "extract-zip": "^2.0.1",
     "fs-extra": "^11.2.0",
     "glob": "11.1.0",
     "rimraf": "^6.1.2",
-    "which": "^2.0.2"
+    "which": "^2.0.2",
+    "yauzl": "^2.10.0"
   },
   "devDependencies": {
     "@swc/core": "^1.15.8",
@@ -53,6 +53,7 @@
     "@types/jest": "^30.0.0",
     "@types/node": "^25.0.3",
     "@types/which": "^3.0.4",
+    "@types/yauzl": "^2.9.1",
     "jest": "^30.2.0",
     "rimraf": "^6.1.2",
     "typescript": "^5.9.3"
